requirements.txt
Flask==3.1.3
mysql-connector-python==9.6.0
Werkzeug==3.1.7
Jinja2==3.1.6
click==8.3.1
itsdangerous==2.2.0
MarkupSafe==3.0.3
blinker==1.9.0
